feat: dropdown menu pluginization
