+++ packages/client/src/built-in/assistant/index.ts
@@ -1,5 +1,6 @@
 import { Plugin } from '../../application/Plugin';
 import { AssistantListProvider, AssistantProvider } from './Assistant.provider';
+import { CalculatorModelProvider } from './calculator/CalculatorModelProvider';
 import { CalculatorProvider } from './calculator/CalculatorProvider';
 import { SearchAndJumpProvider } from './search-and-jump';
 
@@ -9,7 +10,6 @@ export * from './Assistant.provider';
 export class PluginAssistant extends Plugin {
   async load() {
     this.app.use(SearchAndJumpProvider);
-    this.app.use(CalculatorProvider);
     this.app.use(AssistantListProvider, this.options.config);
   }
