+++ packages/client/src/built-in/index.tsx
@@ -27,6 +27,7 @@ import { ACLPlugin } from './acl/ACLPlugin';
 import { AdminLayoutPlugin } from './admin-layout/AdminLayoutPlugin';
 import { WelcomeCard } from './admin-layout/components/WelcomeCard';
 import { PluginAssistant } from './assistant';
+import PluginCalculator from './assistant/calculator/plugin';
 import { AttachmentPreviewPlugin } from './attachment-preview';
 import { PluginBlockSchemaComponent } from './block-schema-component';
 import { PluginContextMenu } from './context-menu';
@@ -418,6 +419,7 @@ export class BuiltInPlugin extends Plugin {
     await this.app.pm.add(CollectionPlugin, { name: 'builtin-collection' });
     await this.app.pm.add(PluginContextMenu, { name: 'context-menu' });
     await this.app.pm.add(PluginAssistant, { name: 'assistant' });
+    await this.app.pm.add(PluginCalculator, { name: 'calculator' });
     await this.app.pm.add(UserSettingsPlugin, { name: 'user-settings' });
     await this.app.pm.add(AttachmentPreviewPlugin, { name: 'attachment-preview' });
     await this.app.pm.add(PluginPageStyle, { name: 'page-style' });
